Add description saving

+++ public/client-js/manage_client.js
@@ -1,3 +1,5 @@
+var timeoutId;
+
 $().ready(function () {
     addPreviewImageEvents();
     $('#cover').click(function () {
@@ -7,6 +9,13 @@ $().ready(function () {
     $('#deleteImage').click(function (e) {
         deleteImage($(this));
         e.stopPropagation();
+    });
+    $('#description').on('input', function() {
+        clearTimeout(timeoutId);
+        timeoutId = setTimeout(function() {
+            // Runs 1 second (1000 ms) after the last change
+            saveDescToDB();
+        }, 1000);
     });
 });
 
@@ -18,6 +27,21 @@ function addPreviewImageEvents() {
         $('#cover').show();
     });
 }
+
+function verticallyCenterImage(){
+    var $img = $('#imageView img'),
+        windowHeight = $(window).outerHeight();
+
+    if($img.height() < windowHeight){
+        var delta = windowHeight - $img.height();
+        $img.css('margin-top', (delta / 2) + 'px');
+    }else{
+        $img.attr('style', '');
+    }
+}
+
+$(window).load(verticallyCenterImage);
+$(window).resize(verticallyCenterImage);
 
 function addImageUrl() {
     var url = encodeURIComponent($('#image_url').val());
@@ -36,8 +60,10 @@ function addImageUrl() {
 }
 
 function deleteImage($img) {
-    alert('確定要刪除圖片？');
-    var imgUrl = encodeURIComponent($('#largeView').attr('src'));
+    if (!confirm('確定要刪除圖片？')) {
+        return;
+    }
+    var imgUrl = encodeURIComponent($('#imageView').children('img').attr('src'));
     $.ajax({
         method: 'DELETE',
         url: '/imageUrl/' + imgUrl
@@ -46,5 +72,42 @@ function deleteImage($img) {
         $('#cover').hide();
         $('#images').html(html);
         addPreviewImageEvents();
+    });
+}
+
+function saveDescToDB() {
+    var savingId;
+    $('#saving').css('color', 'green')
+        .html('儲存中...')
+        .show();
+    savingId = setTimeout(function () {
+        $('#saving').fadeOut(300);
+    }, 2000);
+    $.ajax({
+        method: 'POST',
+        url: '/description',
+        data: {
+            data: $('#description').val()
+        }
+    }).done(function (resp) {
+        clearTimeout(savingId);
+        if (resp == 'OK') {
+            $('#saving').html('已儲存');
+            setTimeout(function () {
+                $('#saving').fadeOut(300);
+            }, 2000);
+        } else {
+            $('#saving').css('color', 'red');
+            $('#saving').html('儲存錯誤！');
+            setTimeout(function () {
+                $('#saving').fadeOut(300);
+            }, 2000);
+        }
+    }).fail(function (err) {
+        $('#saving').css('color', 'red');
+        $('#saving').html('儲存錯誤！');
+        setTimeout(function () {
+            $('#saving').fadeOut(300);
+        }, 2000);
     });
 }
